adds draw visuals and branch
