+++ Application/src/Main.java
@@ -13,6 +13,8 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.ArcType;
 
+import visuals.Draw;
+
 import java.util.Collection;
 
 public class Main extends Application {
@@ -63,10 +65,11 @@ public class Main extends Application {
         pane.add(btn2, 2,        0);
         pane.add(ta, 1, 3, 20, 20);
 
-        Group root = new Group();
-        Canvas canvas = new Canvas(600, 300);
-        GraphicsContext gc = canvas.getGraphicsContext2D();
-        drawShapes(gc);
+
+        Canvas canvas = Draw.newCanvas(600, 300);
+        Draw.drawClass(canvas, 0, 0);
+        Draw.drawClass(canvas, 100, 100);
+
         pane.add(canvas, 22, 3, 20, 20);
 
         StackPane stack = new StackPane();
@@ -77,29 +80,5 @@ public class Main extends Application {
 
         primaryStage.setScene(main);
         primaryStage.show();
-    }
-
-
-    private void drawShapes(GraphicsContext gc) {
-        gc.setFill(Color.GREEN);
-        gc.setStroke(Color.BLUE);
-        gc.setLineWidth(5);
-        gc.strokeLine(40, 10, 10, 40);
-        gc.fillOval(10, 60, 30, 30);
-        gc.strokeOval(60, 60, 30, 30);
-        gc.fillRoundRect(110, 60, 30, 30, 10, 10);
-        gc.strokeRoundRect(160, 60, 30, 30, 10, 10);
-        gc.fillArc(10, 110, 30, 30, 45, 240, ArcType.OPEN);
-        gc.fillArc(60, 110, 30, 30, 45, 240, ArcType.CHORD);
-        gc.fillArc(110, 110, 30, 30, 45, 240, ArcType.ROUND);
-        gc.strokeArc(10, 160, 30, 30, 45, 240, ArcType.OPEN);
-        gc.strokeArc(60, 160, 30, 30, 45, 240, ArcType.CHORD);
-        gc.strokeArc(110, 160, 30, 30, 45, 240, ArcType.ROUND);
-        gc.fillPolygon(new double[]{10, 40, 10, 40},
-                new double[]{210, 210, 240, 240}, 4);
-        gc.strokePolygon(new double[]{60, 90, 60, 90},
-                new double[]{210, 210, 240, 240}, 4);
-        gc.strokePolyline(new double[]{110, 140, 110, 140},
-                new double[]{210, 210, 240, 240}, 4);
     }
 }
